Move tt_prefill_block and tt_prefill_transformer from tt/moe/ to tt/
These modules are not MoE-specific — they compose MLA, FFN/MoE, norm,
embed, and lm_head. Moving them up one level better reflects their role
as top-level prefill building blocks.
